requirements.txt
openai==1.35.13
python-dotenv==1.0.1
networkx=3.3
cdlib=0.4.0
leidenalg==0.10.2
instructor=1.3.4